Move components from lib/components to lib
